Minor bug fixes for duplicate errors and better handling for varied types
I’m conflicted about the second part of this commit. We currently lack
the Number class in Codename One which makes this code awkward… A
better solution might be to put the new value back into the data array
but that would mean that a get operation modifies the data which might
be a problem.